Moved the detect.py to the correct dir
